requirements.txt
django
djangorestframework
selenium
pandas